fix tests for ci and centralize pytest fixtures
