+++ tests/test_upload.py
@@ -1,32 +1,11 @@
-import os
 import time
 from pathlib import Path
 
-import boto3
 import pytest
 from moto import mock_aws
 
 from prescient_sdk.client import PrescientClient
 from prescient_sdk.upload import iter_files, upload
-
-
-@pytest.fixture
-def aws_credentials():
-    """Mocked AWS Credentials for moto."""
-    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
-    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
-    os.environ["AWS_SECURITY_TOKEN"] = "testing"
-    os.environ["AWS_SESSION_TOKEN"] = "testing"
-    os.environ["AWS_DEFAULT_REGION"] = "us-east-1"
-
-
-@pytest.fixture
-def s3(aws_credentials):
-    """
-    Return a mocked S3 client
-    """
-    with mock_aws():
-        yield boto3.client("s3", region_name="us-east-1")
 
 
 @pytest.fixture
